Fixed bug #19188 : Lots of PHP Notices when analyzing the Symfony framework
The list-style… lines were correctly merging their tokens in the CSS tokenizer post-processing but the tokenizer was not going through the curly braces and fixing up the open/close token IDs after removing a token from the array, so those values were invalid.

There is now also an additional check in the PHP tokenizer during post-processing to skip functions without openers. This happens when closures are used inside other function calls. The tokenizer still needs to find some way to parse the code in there, but that's going to be a bit tricky, so muting the notices is fine for now as behaviour doesn't change. Related bug report #19188 is still open for this additional parsing.