Cache files are now stored in the system temp dir (ref #530)
The paths specified for checking and used to locate a common path, which is used to generate the cache file name. When checking a new set of code, if a cache file already exists in a common location, that cache file will be reused and added to instead of a new one being created. So it is possible for a cache file to store the results of multiple projects if you first run over all the projects at once. Running over each individually will still end up finding that first cache file.