Improved detection of STDIN being used.
When using a command like "git diff --name-only | xargs phpcs", files are passed to PHPCS but STDIN is still reported as being available by stream_select(), so the files were ignored. Now, we just read STDIN properly (but non-blocking) and store that cotents for later use. If no files are available, and STDIN is empty, PHPCS will still wait for input on STDIN as it does already.