Changed the way STDIN is read because some strings were stopping after 8096 characters
This wasn't a problem when the stream was blocking, but setting it to non-blocking seems to cause the Sublime Text plugin to break. However it is sending the data on STDIN seems to be different to the way a cat command and pipes do. Reading the content line by line fixes this up, as does adding a sleep in between stream_get_contents calls, but that seems like even more of a hack.