DisallowShortOpenTag: throw warning for potential short open tags in inline HTML
When the ini setting `short_open_tag` is `off` on the machine running PHPCS, short open tags would not be recognized in `T_INLINE_HTML` tokens.
This could lead to undesirable consequences if the code would later be deployed on a machine where `short_open_tag` is set to `on`.

This PR fixes that and will throw a warning if a use of short open tags is detected in inline HTML.

Note: the `getSnippet()` method is a 1-on-1 copy of the one used in the `DisallowAlternativePHPTags` sniff.
It would be better if that method would be placed in another file and just called by both sniff, but I wasn't sure which class would be most suited to place the method in.