Travis tweak: Run the tests with & without the custom ini files.
In the current setup, the tests were only run using the custom ini files.
This meant that the test-cases for which the default ini is expected, were never being run.

I've now added a custom environment variable which toggles whether the default ini values or the custom values should be used.

I've selectively chosen two PHP versions on which to do an extra test run with the custom ini file based on the sniffs involved.
The other PHP versions will now use the default ini values, making the build testing via travis more comprehensive.

For the two extra runs, only the unit tests are run. The other build tests are skipped as they are already covered by the build with the default ini values.