LineLength sniff: add unit tests for existing properties
As the sniff uses the `T_OPEN_TAG` token and then walks through the whole file in one go, the properties need to be changed in the test file **_before_** the open tag is encountered.

In practice, this means that the unit tests for the changed properties need to be in separate files.