Squiz/EmbeddedPHP sniff - prevent fixer conflict
The `Squiz.PHP.EmbeddedPhp` sniff will create a fixer conflict with itself when it encounters a `//` comment just before the PHP close tag.

The tokenizer will tokenize whitespace at the end of a `//` comment as part of the `T_COMMENT` token, so the fixer just keeps adding more spaces between the `T_COMMENT` and the `T_CLOSE_TAG` which in the next fixer round is then tokenized again as being part of the `T_COMMENT` resulting in the loop.