:sparkles: Add new `AssignmentInCondition` sniff to `Generic`
Most of the time variable assignments being made in conditions are unintentional and are in actual fact typos which were intended as a comparison.

This sniff will warn when any such a variable assignment is found in a condition.

Note: the sniff does currently **not** detect variable assignments in the conditional part of ternaries.

Included extensive unit tests.