DisallowTabIndent: Use the correct error message and code
The "TabsUsed" error message and code were only set before the loop and not reset within the loop.

In practice this meant that as soon as the first non-indent error is encountered, the `$error` and `$errorCode` would be overruled and for any indent errors found after that, the wrong message and error code would be thrown.