:sparkles: New ArbitraryParenthesesSpacing sniff for Core
Checks the spacing on the inside of arbitrary parentheses.

Arbitrary parentheses, for the purposes of this sniff, are all parentheses which do not belong to a control structure, function declaration, function call or language construct.
This is to prevent conflicts and duplication with rules for those type of constructs.

The sniff will throw (fixable) errors when an incorrect amount of whitespace (too much/too little) has been found on the inside or the parenthesis.
The sniff will throw a warning when an empty set of arbitrary parentheses is encountered.

Includes extensive unit tests.

- [ ] Add info on the custom properties to the wiki