Squiz/InlineComment: fix fixer conflict when comment found at end of function
When an inline comment is found at the end of a function, removing a blank line after it conflicts with the `Squiz.WhiteSpace.FunctionClosingBraceSpace` sniff which demands a blank line at the end of a function.

I have chosen to fix this by adding a specific error code for that situation and excluding that error code from the `Squiz` ruleset.

This way, the backward-compatibility break for other standards using the `Squiz.Commenting.InlineComment` sniff will be smallest.
Only if a standard explicitly in/excluded the `Squiz.Commenting.InlineComment.SpacingAfter` errorcode will this have any effect on them.

The unit tests already contain a test covering this. As the errorcode is excluded via the ruleset, the results of the unit tests do not change.

Fixes 1709