Squiz/MemberVarSpacing: improve handling of multi-line multi-property declarations
The sniff did not handle multi-property declarations well.
Up to now, it would:
* correct the number of lines above the line for each variable
* not correct the number of lines above the line containing the scope keyword

IMHO this is incorrect and only the lines above the line containing the scope keyword should be checked & fixed.

This has now been fixed.

By returning the `$stackPtr` of the semi-colon at the end of the property declaration, we skip past the additional variables.

Includes unit tests.