Squiz/ClassDeclaration: bug fix - ignore PHPCS annotation tokens
Treat the new `// phpcs:` comments in the same way as "ordinary" `T_COMMENT` tokens.

Includes adjusted unit test.

This fixes a bug where an error would be reported for a `// phpcs:` comment being on the same line as the class close brace, as well as the fixer for that error moving the `// phpcs:` comment, while it shouldn't.