Squiz/ControlStructureSpacing: bug fix - allow for PHPCS annotation tokens
Treat the new `// phpcs:` comments in the same way as "ordinary" `T_COMMENT` tokens.

Includes unit test.

Related to 1817, which was recently fixed in a7d4118b9ca880acd2b9a7d70b7861e9228756eb. This commit builds onto that fix.

Without the fix, the `NoLineAfterClose` error would not be reported when there is a `phpcs:` comment after a control structure close brace and the needed fix would not be applied.