Squiz/NonExecutableCode: split the test file into two
The test file for the NonExecutableCode sniff was very unwieldy as it was as the last 50 or so lines were one unit test which had to be at the end of the file, which meant that any new unit tests to be added would have to be added above that and the return array in `getWarningList()`  would need to be renumbered.

This commit splits the last unit test off into a separate unit test file to make life easier.