Generic/MultipleStatementAlignment: fix bug/list syntax before assignment operator
This simplifies the fixes which were previously committed to fix 1870 and 1848, which were released in `3.2.3`.

Both those fixes try to prevent the sniff getting confused over (potentially) multi-line value assignments, such as closures and arrays.

The simplification removes the previously added code in favour of skipping to the end of a statement as soon as an assignment operator has been found.

This might also make the sniff marginally faster as the tokens between the assignment operator and the end of the statement will no longer be examined.

N.B.: The reason for going one back from the end of the assignment, is that otherwise the code addressing the unit test on line 220 / added in commit 579d39cac3163bdf05ce1b50a7e470acd452735b would not be triggered, leading to incorrect results.

Side-note: I think it would be a good idea to add unit tests for the `File::findEndOfStatement()` method.