Further fix for #1925
Performance was terrible for files with heaps of closures + the array skipping code meant that closures used inside array values were being skipped as well. Now, instead of looking for closures and anon classes, we check token levels and process assignments inside nested blocks before returning to the main block. Blank lines are skipped inside arrays by keeping track of if we are in one.