Generic/OpeningBrace sniffs: fix bug in fixer handling of PHPCS annotations
As discussed in 1931, moving a trailing `phpcs:ignore` annotation to the next line changes its meaning, which is what currently happens when the `ContentAfterBrace` error is triggered by one of the new PHPCS annotations in the `Generic.Functions.OpeningFunctionBraceKernighanRitchie` and `Generic.Functions.OpeningFunctionBraceBsdAllman` sniffs.

For the `Generic.Functions.OpeningFunctionBraceBsdAllman` sniff, this also impacts the fixer of the `BraceOnSameLine` errror which could also move the comment down.

This PR fixes that.

For this fix, I've chosen to ignore *all* the PHPCS annotations, not just the `phpcs:ignore` one. If so desired, I can change the PR to only make an exception for a trailing `phpcs:ignore` comment and to move `phpcs:disable/enable/set` comments to the next line, just like "normal" comments.

Includes unit tests.