Replace slow Tokens::tokenName() with define()
I actually profiled this and found Tokens::tokenName() is called hundreds of thousands of times here. Even if the method implementation appears to be trivial, it is expensive. The line touched in this patch always converts the same integer constant to a string, and compares the string then. In the profile run I did this line alone accounts for 3% of the total runtime. PHPs define() is much faster, even if called as often as it is the case here.