:sparkles: New Generic.WhiteSpace.IncrementDecrementSpacing sniff
New sniff to check that there are no spaces (or comments) between an incrementor/decrementor and the variable it applies to.

The sniff can examine both PHP as well as JS files for this, though I suspect there may be some false positive reports for JS at some point as `T_STRING` is quite a generic token to check for.

Includes unit tests.
Includes auto-fixer.
Includes code-consistency metrics.

I've made a conscious decision to not make the number of spaces accepted between the in/decrementor and the variable configurable as it is a widely accepted convention that there should be no spaces between them.
If it would be so desired, I can add a public property to make this configurable.

Fixes 2172