Squiz/MethodScope: examine anonymous classes without throwing double errors
Related to the test case provided in 2178.

The `Squiz.Scope.MethodScope` sniff did not examine anonymous classes and when those where added to the scope, the sniff would still ignore issues in nested classes as anything nested in a function would be ignored.

The reasoning and solution used is the same as previously applied in 2182 for the `Generic.NamingConventions.ConstructorName` sniff.

> Turns out the `processTokensWithinScope()` method gets called for each whitelisted scope a `T_FUNCTION` token exists in.
>
> In effect, this means that the first time it gets called for the `nested_function()` method, it is because it is within the scope of the `Nested` class, the second time the function gets called, it is for the method being in the scope of the anonymous class.
>
> The fix I've now applied checks that the scope for which the method is called is actually the deepest scope first and bows out if it's not.

This fixes the error messages not being thrown.

Includes unit tests.