File::getMemberProperties(): don't recognize method params as properties
Method parameters were recognized as if they were class properties.

Looks like I already included an initial unit test case for this in 1715, but that case was not being tested.
The change in this PR should fix the issue and activates the existing unit test and adds some extra as well.