PSR2/EndFileNewline: add missing fixed files + minor improvement
The sniff contains fixers, but didn't have `fixed` files.

Additionally, I've hard-coded the end of line character to be used by the fixer to `\n` as PSR2 demands that anyway.
This should give a minor efficiency improvement for when a file has, for instance, Windows line endings which would be recorded as such in `$phpcsFile->eolChar`. This way the `Generic.Files.LineEndings` sniff doesn't have to go over the file to correct it (again) in a second loop.