Squiz/PostStatementComment: fix fixer conflict
[Fixer conflicts series PR]

When a JS file contains a `//` comment on the very first line of the file, the fixer unnecessarily added a new line before it, which would cause a fixer conflict with the `Squiz.WhiteSpace.SuperfluousWhitespace` sniff.

The fixer conflict can be reproduced by checking out just the new `PostStatementCommentUnitTest.2.js` test case file from this PR and running the sniff - as it is currently in `master` - over it using the `Squiz` standard.

`phpcbf -s ./Squiz/Tests/Commenting/PostStatementCommentUnitTest.2.js --standard=Squiz -vv`

This commit fixes the fixer conflict and adds a new unit test file guarding against it.

Additional notes:
* Added minor defensive coding improvements.
* Added tests confirming the correct handling of post statement comments for PHP closures.
* Added tests confirming the correct handling of the new PHPCS annotations.
* Removed some redundant break statements from the unit test file.