PSR2/PSR12 rulesets: prevent duplicate messages
The `PSR2.ControlStructures.ControlStructureSpacing` sniff already checks, throws errors and fixes spacing after the opening parenthesis and closing parenthesis of a `for()` statement.

As the `Squiz.ControlStructures.ForLoopDeclaration` sniff does so as well, this leads to two errors for the same issue being thrown with slightly different messages.

This tweaks prevents this message duplication.