Squiz/ForLoopDeclaration: improve the accuracy of the SpaceAfterOpen/SpaceBeforeClose error messages
The error messages for `SpaceAfterOpen`/`SpaceBeforeClose` did not handle situations where the spacing found was a new line.

The determination of `$spaceAfterOpen` and `$spaceBeforeClose` has now been brought in line with other sniffs and will report `newline` when a new line was found.

Includes using the predetermined `length` of the token instead of doing length calculations.

For the "spaces found, no spaces allowed" situations, only a minor tweak has been made to the error message to acknowledge that more than just a space may have been found.