Squiz/ForLoopDeclaration: refactor semi-colon related code
The first and second semi-colon examinations were basically the same, i.e. duplicate code.

This refactors the semicolon examination to use a `do-while` loop on the semicolons within a `for`, removing the code duplication (and opening the way for further improvements in the next few commits).

This commit does not change the behaviour of the sniff.