Squiz/ForLoopDeclaration: bugfix with closures being used in expressions
When a closure would be used within an expression, the code within the closure would also contain semicolons.

The existing code did not take that into account and would throw false positives for spacing around semicolons in the closures (not the concern of this sniff) and would incorrectly neglect to throw errors for spacing around semicolons after the closure.

This is now fixed.