Remove unnecessary strlen() function calls
The `$tokens` array already contains the `length` of each and every token, so there is no need to (re-)calculate this in individual sniffs, unless we're examining `T_COMMENT` or `T_INLINE_HTML` tokens.

Using the `length` from the tokenizer array should also be more accurate when tabs have been replaced with spaces or when dealing with Unicode texts.