Squiz/FunctionCommentThrowTag: fix incorrect `Missing` errors with FQN Exception names
The code did not take into account that Fully Qualified names could be used in the `@throws` tag while a relative Exception name may be used in the `throw` statement.

Includes unit tests (probably a few too many, but I wanted to be sure that I understood how the sniff was dealing with this).