PSR12/ShortFormTypeKeywords: bug fix / false positive
The PSR 12 standard (currently) says nothing about whether or not spaces within type casts are allowed.

As it was, the `PSR12.Keywords.ShortFormTypeKeywords` sniff would throw a false positive for a short form type cast like `( bool )` and "fix" it by removing the spaces, which is not the concern of this sniff.

If at some point the PSR12 standard would take a stand about spaces within type casts, the `Squiz.WhiteSpace.CastSpacing` sniff could be added to check & fix these issues.

In the mean time, the `ShortFormTypeKeywords` sniff should be space-agnostic.

This PR takes care of that.

It also makes the error message slightly more descriptive by being explicit about the type cast found in the error message.

Includes unit tests.