Hotfix for PEAR install
In the WordPressCS  repo we received a bug report from someone using PHPCS 3.4.0 but missing the `Generic.WhiteSpace.LanguageConstructSpacing` sniff.

Turned out that that sniff is missing from the `package.xml` file which is why it's not available when someone is using a PEAR install of PHPCS.

I've now done a comprehensive check of all relevant files in the repo versus the files listed in the `<contents>` tag of the `package.xml` file and found some more discrepancies with this commit as a result.

As this basically means that the PEAR install is currently "broken" with regards to the `Code` report (PHPCS 3.0), the `Generic.Formatting.SpaceBeforeCast` sniff (PHPCS 3.4.0) as well as the `Generic.WhiteSpace.LanguageConstructSpacing` sniff (PHPCS 3.3.0), this should be considered a hotfix.