Use Exceptions consistently
As far as I can see, Exceptions thrown by sniffs/utility methods when, for instance, a wrong token is passed, should be `RuntimeException`s, while Exceptions thrown by the `Tokenizer` classes should be `TokenizerException`s.

In a number of places this was applied inconsistently. This has now been fixed.

Aside from that, for quite a number of methods, the `Exception` being thrown has now been documented.