Generic/DeprecatedFunctions: add missing unit tests
[Sniff completeness series PR]

Note: PHP internally doesn't always correctly register functions as deprecated, so this sniff will only work when it has (as the information is otherwise not available to the Reflection extension).

Some trial and error showed that the two functions now in the tests are correctly registered, but numerous other functions I tried to add to the test case file failed as PHP didn't provide the expected information.