Generic/InlineControlStructure: bug fix
Prevent creating incorrect fixes and getting into a fixer conflict when both short and long open tags are used in combination with alternative control structures and the short open tags are not recognized because the system on which the scan is run has `short_open_tag=Off`.

Solved by bowing out early when an alternative control structure is detected which was not recognized as such (`scope_opener` not set).

Fixes 2565